@@ -23,12 +23,6 @@ import android.content.res.Resources;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
 import android.os.Handler;
-import android.support.annotation.ColorRes;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.NonNull;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.content.res.ResourcesCompat;
-import android.support.v4.graphics.drawable.DrawableCompat;
 import android.util.DisplayMetrics;
 import android.util.TypedValue;
 import android.view.Display;
@@ -40,6 +34,13 @@ import android.widget.EditText;
 import android.widget.ImageView;
 
 import java.util.Locale;
+
+import androidx.annotation.ColorRes;
+import androidx.annotation.DrawableRes;
+import androidx.annotation.NonNull;
+import androidx.core.content.ContextCompat;
+import androidx.core.content.res.ResourcesCompat;
+import androidx.core.graphics.drawable.DrawableCompat;
 
 public class Util {
 
@@ -116,7 +117,7 @@ public class Util {
      *
      * @param resId the resource id for the given drawable
      * @return a wrapped drawable ready fo use
-     * with {@link android.support.v4.graphics.drawable.DrawableCompat}'s tinting methods
+     * with {@link androidx.core.graphics.drawable.DrawableCompat}'s tinting methods
      * @throws Resources.NotFoundException
      */
     public static Drawable getWrappedDrawable(Context context, @DrawableRes int resId) throws Resources.NotFoundException {
