@@ -16,8 +16,9 @@ package com.arlib.floatingsearchview.util.adapter;
  * limitations under the License.
  */
 
-import android.support.v7.widget.RecyclerView;
 import android.view.MotionEvent;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 public abstract class OnItemTouchListenerAdapter implements RecyclerView.OnItemTouchListener {
 
