@@ -18,8 +18,6 @@ package com.arlib.floatingsearchviewdemo.adapter;
 
 import android.app.Activity;
 import android.graphics.Color;
-import android.support.v7.graphics.Palette;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -32,6 +30,8 @@ import com.arlib.floatingsearchviewdemo.data.ColorWrapper;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 public class SearchResultsListAdapter extends RecyclerView.Adapter<SearchResultsListAdapter.ViewHolder> {
 
