@@ -8,8 +8,6 @@ import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.content.res.ResourcesCompat;
 import android.text.Html;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -29,6 +27,9 @@ import com.arlib.floatingsearchviewdemo.data.ColorSuggestion;
 import com.arlib.floatingsearchviewdemo.data.DataHelper;
 
 import java.util.List;
+
+import androidx.annotation.Nullable;
+import androidx.core.content.res.ResourcesCompat;
 
 
 public class SlidingSearchViewExampleFragment extends BaseExampleFragment {
